// admin/src/api/generated/schemas/adminAuditLogItem.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Internal Wallet Admin API
 * Admin API (served via api-gateway)
 * OpenAPI spec version: v1
 */
import type { AdminOperatorInfo } from './adminOperatorInfo';
import type { AdminAuditLogItemDetails } from './adminAuditLogItemDetails';

export type AdminAuditLogItem = {
  id?: string;
  operator?: AdminOperatorInfo;
  action?: string;
  target_type?: string;
  target_id?: string;
  description?: string;
  details?: AdminAuditLogItemDetails;
  ip?: string;
  user_agent?: string;
  created_at?: string;
  success?: boolean;
  grpc_code?: number;
  error_message?: string;
  request_id?: string;
  rpc_method?: string;
  http_method?: string;
  http_path?: string;
  duration_ms?: number;
  operator_email?: string;
  module?: string;
  idempotency_key?: string;
};